Correctly record history of returns_async return values.
Prior to this commit return values defined using the returns_async
keyword did not properly appear in the history. This is because
returns_async was not considered when creating the list of
expansions.

Once this was added, there was a second issue because the old
void/callback expansion was seen after the promise-based one
and overwrote the history data we had generated. To avoid this
I added logic which always prefers a return type over void.