Correctly determine channel for APIs with dependencies.
When considering dependencies, we were incorrectly choosing the
"most released" channel as the channel an API became available.
This was causing some issues as, for example, action.openPopup
is still restricted to dev but it has a dependency of the action
key in the manifest which has existed in stable for a while.

Fixing this should make the openPopup API correctly appear as
restricted to dev, and also fix some other similar issues across
the generated types.