Prefer features without `location` condition
The action.openPopup() API is available to policy installed
extensions, OR to any extension as long as a (default enabled)
feature flag is set. We can't render complex unions like this so
instead just prefer the result without the `location` condition.

The result of this is that openPopup() will show as available on
stable without the "Requires policy" badge / TypeScript comment.