Fix permissions for management API
A recent Chromium change [1] exposed the management API to the Chrome
Web Store for testing purposes. Unfortunately, this meant chrome-types
was thinking that the API was available without the `management`
permission.

Unfortunately, I couldn't see an easy way to fix this.

To resolve that, I've made two tweaks here as workarounds:

1) Ignore feature definitions that are only for webpages. We don't
really ever have a reason to show those in our documentation.

2) Teach the tool that if no channel is specified, this is equivalent to
an API being available on stable.

Together, these seem to be enough to cause the tool to filter out the
wrong definitions and pick up the ones used by most developers.

[1] https://source.chromium.org/chromium/chromium/src/+/59d0c94348743ce1831a113b81fd5dc420d68919:extensions/common/api/_api_features.json;dlc=30d48f6ef398e61601b43bcea809ac846fb99535