Disable compression to reduce CI failures
There have been a number of recent CI failures [1] which are related to
zlib compression when downloading files from the Chromium repository.
This may be an unrelated issue (e.g throttling / network instability)
but disabling compression to see if it helps. I was able to get at least
one CI pass using this [2].

[1] https://github.com/GoogleChrome/chrome-types/actions/runs/15348275531
[2] https://github.com/GoogleChrome/chrome-types/actions/runs/15348791346