Add mechanism to manually patch history for DevTools APIs.
We are used to handling cases where an API was incorrectly marked as
`nodoc`. In those situations, we can use override.js to ignore it.

DevTools is a special case because the schema is not used for anything,
so APIs can exist in stable but be missing entirely from the schema. Add
a mechanism to override the history when needed.

This replaces the previous mechanism to do this which was more limited
as you were unable to set a new low version.